Add files via upload
All of the code for the project. The images are for the Gui, the Read code is to program the NFC tags to be a certain string, the test drawers is the code for the servo motor functions, and the ProjectGui.py is the code you need to run for the project to work.